# [Performance Improvement] Pool `gzip.Writer` and `bytes.Buffer`