--- a/elastictransport/elastictransport.go
+++ b/elastictransport/elastictransport.go
@@ -105,6 +105,12 @@ type Config struct {
 	CertificateFingerprint string
 }
 
+type gzipWriter struct {
+	writer *gzip.Writer
+	buf    *bytes.Buffer
+	err    error
+}
+
 // Client represents the HTTP client.
 type Client struct {
 	sync.Mutex
@@ -131,6 +137,7 @@ type Client struct {
 
 	compressRequestBody      bool
 	compressRequestBodyLevel int
+	gzipWriterPool           sync.Pool
 
 	instrumentation Instrumentation
 
@@ -269,6 +276,18 @@ func New(cfg Config) (*Client, error) {
 		client.compressRequestBodyLevel = gzip.DefaultCompression
 	}
 
+	client.gzipWriterPool = sync.Pool{
+		New: func() any {
+			buf := new(bytes.Buffer)
+			writer, err := gzip.NewWriterLevel(buf, client.compressRequestBodyLevel)
+			return &gzipWriter{
+				writer: writer,
+				buf:    buf,
+				err:    err,
+			}
+		},
+	}
+
 	return &client, nil
 }
 
@@ -292,27 +311,27 @@ func (c *Client) Perform(req *http.Request) (*http.Response, error) {
 
 	if req.Body != nil && req.Body != http.NoBody {
 		if c.compressRequestBody {
-			var buf bytes.Buffer
-			zw, err := gzip.NewWriterLevel(&buf, c.compressRequestBodyLevel)
-			if err != nil {
-				fmt.Errorf("failed setting up up compress request body (level %d): %s",
+			gzipWriter := c.newGzipWriter()
+			if gzipWriter.err != nil {
+				return nil, fmt.Errorf("failed setting up up compress request body (level %d): %s",
 					c.compressRequestBodyLevel, err)
 			}
-			if _, err = io.Copy(zw, req.Body); err != nil {
+			defer c.gzipWriterPool.Put(gzipWriter)
+
+			if _, err = io.Copy(gzipWriter.writer, req.Body); err != nil {
 				return nil, fmt.Errorf("failed to compress request body: %s", err)
 			}
-			if err = zw.Close(); err != nil {
+			if err = gzipWriter.writer.Close(); err != nil {
 				return nil, fmt.Errorf("failed to compress request body (during close): %s", err)
 			}
 
 			req.GetBody = func() (io.ReadCloser, error) {
-				r := buf
-				return ioutil.NopCloser(&r), nil
+				return ioutil.NopCloser(gzipWriter.buf), nil
 			}
 			req.Body, _ = req.GetBody()
 
 			req.Header.Set("Content-Encoding", "gzip")
-			req.ContentLength = int64(buf.Len())
+			req.ContentLength = int64(gzipWriter.buf.Len())
 
 		} else if req.GetBody == nil {
 			if !c.disableRetry || (c.logger != nil && c.logger.RequestBodyEnabled()) {
@@ -558,4 +577,15 @@ func (c *Client) logRoundTrip(
 		}
 	}
 	c.logger.LogRoundTrip(req, &dupRes, err, start, dur) // errcheck exclude
+}
+
+func (c *Client) newGzipWriter() *gzipWriter {
+	gzipWriter := c.gzipWriterPool.Get().(*gzipWriter)
+	if gzipWriter.err != nil {
+		return gzipWriter
+	}
+
+	gzipWriter.buf.Reset()
+	gzipWriter.writer.Reset(gzipWriter.buf)
+	return gzipWriter
 }