--- a/elastictransport/elastictransport_benchmark_test.go
+++ b/elastictransport/elastictransport_benchmark_test.go
@@ -22,6 +22,7 @@ package elastictransport_test
 
 import (
 	"io/ioutil"
+	"math/rand"
 	"net/http"
 	"net/url"
 	"strings"
@@ -86,4 +87,53 @@ func BenchmarkTransport(b *testing.B) {
 			}
 		}
 	})
+
+	body := strings.NewReader(generateRandomString(100))
+	b.Run("Compress body (bodysize: small)", func(b *testing.B) {
+		for i := 0; i < b.N; i++ {
+			tp, _ := elastictransport.New(elastictransport.Config{
+				URLs:                []*url.URL{{Scheme: "http", Host: "foo"}},
+				Transport:           newFakeTransport(b),
+				CompressRequestBody: true,
+			})
+
+			req, _ := http.NewRequest("GET", "/abc", body)
+			_, err := tp.Perform(req)
+			if err != nil {
+				b.Fatalf("Unexpected error: %s", err)
+			}
+		}
+	})
+
+	body = strings.NewReader(generateRandomString(1_000_000))
+	b.Run("Compress body (bodysize: large)", func(b *testing.B) {
+		for i := 0; i < b.N; i++ {
+			tp, _ := elastictransport.New(elastictransport.Config{
+				URLs:                []*url.URL{{Scheme: "http", Host: "foo"}},
+				Transport:           newFakeTransport(b),
+				CompressRequestBody: true,
+			})
+
+			body := strings.NewReader(`FAKE`)
+
+			req, _ := http.NewRequest("GET", "/abc", body)
+			_, err := tp.Perform(req)
+			if err != nil {
+				b.Fatalf("Unexpected error: %s", err)
+			}
+		}
+	})
+}
+
+func generateRandomString(length int) string {
+	const charset = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
+	result := make([]byte, length)
+
+	seed := rand.NewSource(1)
+	random := rand.New(seed)
+
+	for i := range result {
+		result[i] = charset[random.Intn(len(charset))]
+	}
+	return string(result)
 }