## Summary
Added `sync.Pool` to pool `gzip.Writer` and `bytes.Buffer` for better performance.

## Background
While I was investigating profile report of my application, noticed transport client is not pooling gzip writer and buffer.
As discussed in [this Reddit thread](https://www.reddit.com/r/golang/comments/9uejp4/usage_of_syncpool_for_gzipwriter_in_http_handlers/) and many other places, pooling them will benefit us by reusing allocated memory space rather than allocating and freeing them on every request.

## Performance Comparison
These are the benchmark result on my machine (Intel(R) Core(TM) i5-8259U CPU @ 2.30GHz)

### Compress body 
|                        | ns/op  | B/op   | allocs/op |
| ---------------------- | ------ | ------ | --------- |
| Before                 | 129744 | 815064 | 32        |
| After(Without pooling) | 117462 | 815014 | 31        |
| After(With poolong)    | 20462  | 1152   | 12        |

bench result
```
Running tool: /usr/local/bin/go test -benchmem -run=^$ -bench ^BenchmarkTransport$ github.com/elastic/elastic-transport-go/v8/elastictransport

goos: darwin
goarch: amd64
pkg: github.com/elastic/elastic-transport-go/v8/elastictransport
cpu: Intel(R) Core(TM) i5-8259U CPU @ 2.30GHz

=== RUN   BenchmarkTransport/Compress_body_(pool:_false)
BenchmarkTransport/Compress_body_(pool:_false)
BenchmarkTransport/Compress_body_(pool:_false)-8                    9450            117462 ns/op          815014 B/op         31 allocs/op
=== RUN   BenchmarkTransport/Compress_body_(pool:_true)
BenchmarkTransport/Compress_body_(pool:_true)
BenchmarkTransport/Compress_body_(pool:_true)-8                    57690             20462 ns/op            1152 B/op         12 allocs/op

====================
=== before the change ===
====================

=== RUN   BenchmarkTransport/Compress_body
BenchmarkTransport/Compress_body
BenchmarkTransport/Compress_body-8                    9697            129744 ns/op          815064 B/op         32 allocs/op
```