@pakio 2024-02-29
Hi @Anaethelion, would you mind reviewing this PR when you are available?
I've been using this library via go-elasticsearch in production and this change would really help our application run more stable with high traffic volume environment.