@pakio 2024-03-10
hi @Anaethelion , thanks for your feedback!

> While this seems like a worthy improvement for some use cases, it turns out that this could hurt a lot others and create some contention around the pool which could lead to OOM.

Understood. Just to clarify, does this mean you've actually experienced OOM with sync.Pool or with this implementation, or have general concern about it?

>  however I'm interested to know if you have more examples of real world use cases beyond the benchmark

sure, here are some examples of use cases of `sync.Pool` along with `gzip.NewWriter`

grpc-go: [link](https://github.com/grpc/grpc-go/blob/7c377708dc070bc6f4dfeedfb3c3f38c3410912b/encoding/gzip/gzip.go)
prometheus golang client: [link](https://github.com/prometheus/client_golang/blob/b59cfc9f6e7f8ce4cc855be54f05fcf58a48b247/prometheus/promhttp/http.go#L57-L61)

And our team have also been using the feature on the platform for more than a year and is operating without any problem.

> How does it help you application being more stable ? What kind of usage do you have where this helps ? Search or Bulk indexing?

Our team uses [go-elasticsearch](https://github.com/elastic/go-elasticsearch) for both data indexing pipeline and search application. 
While I was exploring the CPU time profile graph of our search application, interestingly more than 5% of the time were consumed on `elastictransport.Perform`, specifically to [allocate memory for copying bytes to buffers](https://github.com/elastic/elastic-transport-go/blob/main/elastictransport/elastictransport.go#L301-L303) . The size of requests we sending is generally around 3~5KB. Unfortunately we don't have profile for indexing pipeline but using [esutil.BulkIndexer](https://github.com/elastic/go-elasticsearch/blob/ab9e8b523888d41d85d475e5440c8634ccf84be7/esutil/bulk_indexer.go#L38) to batch-send requests with threshold of 10MB, therefore I suspect the impact would be larger.

> We still have the option of maintaining the current behavior and adding the pool as an optional behavior.

I'm happy to make this change as well as adding more practical test case using actual json file, but first would like to understand your concern about OOM by using `sync.Pool`.