@Anaethelion 2024-03-11
> Understood. Just to clarify, does this mean you've actually experienced OOM with sync.Pool or with this implementation, or have general concern about it?

This is just general concern around edge cases while pushing the `BulkIndexer` within its limits, I have a feeling this could lead to memory exhaustion if the machine that runs the indexing tries to do other tasks and the pool has grown too big.
My point is, it could break existing usages and I'd like not to do that. :)

> sure, here are some examples of use cases of `sync.Pool` along with `gzip.NewWriter`
> 
Thank you!

> Our team uses [go-elasticsearch](https://github.com/elastic/go-elasticsearch) for both data indexing pipeline and search application. While I was exploring the CPU time profile graph of our search application, interestingly more than 5% of the time were consumed on `elastictransport.Perform`, specifically to [allocate memory for copying bytes to buffers](https://github.com/elastic/elastic-transport-go/blob/main/elastictransport/elastictransport.go#L301-L303) . The size of requests we sending is generally around 3~5KB. Unfortunately we don't have profile for indexing pipeline but using [esutil.BulkIndexer](https://github.com/elastic/go-elasticsearch/blob/ab9e8b523888d41d85d475e5440c8634ccf84be7/esutil/bulk_indexer.go#L38) to batch-send requests with threshold of 10MB, therefore I suspect the impact would be larger.

Got you. It matches what I've benchmarked locally as well. The greater the payload, the greater the gain which makes sense. At the same time for search heavy use cases I've found this could slow down the client depending on the compression that has been set up.

> I'm happy to make this change as well as adding more practical test case using actual json file, but firstly I would like to understand your concern regarding OOM by using sync.Pool so that I can properly address the issue.

That concern only led me to consider making this more pluggable at least at first and then in time or for a new major we can flip the switch and make that the default. But as much as we can I'd like not to break existing behavior.