@pakio 2024-03-12
> I have a feeling this could lead to memory exhaustion if the machine that runs the indexing tries to do other tasks and the pool has grown too big.

It sounds a fair concern. 
Will update the PR soon to make it optional, and will ping you when it's ready!