@pakio 2024-03-12
I've made a few changes on my implementation
- Moved gzip compression logic to separate file
- Added `gzipCompressor` interface, added `simpleGzipCompressor` and `pooledGzipCompressor` inherits the interface
- Added `Config.PoolCompressor` to determine which compressor to use from client

I also found a bug in the benchmark that 1. reader was always at the end of the cursor except first loop 2. tp was always recreated in the bench which means the pool is not utilized properly. I updated them as well as the benchmark result in the description.

ptal when you get a chance!